Redifine model methods for BookOps view book, view all and delete
